refactor: 플레이리스트 저장 공통로직 훅 생성
