+++ components/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./useToast";
+export * from "./useSavePlaylist";
